// frontend/src/lib/api-client/models/LeaderboardEntry.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LeaderboardEntry = {
    rank?: number;
    userId?: string;
    username?: string;
    avatar?: string;
    sessionsCompleted?: number;
    avgScore?: number;
    totalTime?: string;
};